requirements.txt
black==25.1.0
boto3==1.35.30
botocore==1.35.99
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.8.28
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.191
ultralytics-thop==2.0.16
urllib3==2.5.0
